pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "oxapy==0.7.3",
    "pyproject-toml>=0.1.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy-libsql==0.2.0",
]

[tool.setuptools]
packages = ["core"]
